Add ``pgclone copy`` command.
The ``pgclone copy`` command is a shortcut for running ``CREATE DATABASE <target> TEMPLATE <source>``
for doing quick copies. This command complements local ``pgclone restore`` commands.

For example, copy the current database with ``pgclone copy``, and quickly restore it with
``pgclone restore :current``. Use a custom name with ``pgclone copy :custom_name`` and restore it
with ``pgclone restore :custom_name``.

Note that this command takes out an exclusive lock on the source database, meaning it should not be
executed in production environments.

Type: feature